chore(quicping): moved quic logic into separate file
